src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/svg-style-in-shadow-tree-crash-expected.txt

CONSOLE ERROR: line 9: Uncaught TypeError: document.registerElement is not a function
CONSOLE ERROR: line 9: Uncaught TypeError: document.registerElement is not a function
CONSOLE ERROR: line 9: Uncaught TypeError: document.registerElement is not a function
CONSOLE ERROR: line 9: Uncaught TypeError: document.registerElement is not a function
CONSOLE ERROR: line 9: Uncaught TypeError: document.registerElement is not a function
CONSOLE ERROR: line 9: Uncaught TypeError: document.registerElement is not a function
CONSOLE ERROR: line 9: Uncaught TypeError: document.registerElement is not a function
CONSOLE ERROR: line 9: Uncaught TypeError: document.registerElement is not a function
CONSOLE ERROR: line 9: Uncaught TypeError: document.registerElement is not a function
CONSOLE ERROR: line 9: Uncaught TypeError: document.registerElement is not a function
CONSOLE ERROR: line 9: Uncaught TypeError: document.registerElement is not a function
PASS
